requirements.txt
django
pillow
wheel
django-sslserver
django-extensions
kombu
pyOpenSSL
werkzeug
wfastcgi
django_q
django-simple-history
xhtml2pdf
django-email
pandas
numpy
fpdf
pytest-shutil
matplotlib
daphne
channels
channels-redis